+++ game/src/models/Rumba.h
@@ -7,11 +7,11 @@
 
 class Rumba : public Object {
 	protected:
-		Vector<int> center_pos;
+		Vector<float> center_pos;
 		int radius;
 	public:
-		Rumba(int x, int y, int r) : center_pos(Vector<int>(x, y)), radius(r) {}
-		Vector<int> getCenterPos() { return center_pos; }
+		Rumba(int x, int y, int r) : center_pos(Vector<float>(x, y)), radius(r) {}
+		Vector<float> getCenterPos() { return center_pos; }
 		int getRadius() { return radius; }
 };
